# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HomeBase API",
    "description": "Authentication, user profile, catalog reference data, and listings discovery endpoints for the HomeBase mobile app.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:4000", "description": "Local development" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from login, register, refresh, or social auth"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["message", "status"],
            "properties": {
              "message": { "type": "string", "example": "Invalid credentials" },
              "status": { "type": "integer", "example": 401 }
            }
          }
        }
      },
...
      "ListingPhoto": {
        "type": "object",
        "required": ["id", "cloudinaryPublicId", "url", "position"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "cloudinaryPublicId": { "type": "string", "example": "listings/abc123/photo1" },
          "url": { "type": "string", "format": "uri", "example": "https://res.cloudinary.com/demo/image/upload/v1/listings/abc123/photo1.jpg" },
          "position": { "type": "integer", "example": 0 }
        }
      },
      "RentDetails": {
        "type": "object",
        "properties": {
          "monthlyRent": { "type": "integer", "nullable": true, "description": "Amount in kobo", "example": 85000000 },
          "annualRent": { "type": "integer", "nullable": true, "description": "Amount in kobo", "example": 960000000 },
          "securityDeposit": { "type": "integer", "example": 85000000 },
          "leaseTermMonths": { "type": "integer", "example": 12 },
          "availableFrom": { "type": "string", "nullable": true, "format": "date-time" }
        }
      },
      "SaleDetails": {
        "type": "object",
        "required": ["salePrice"],
        "properties": {
          "salePrice": { "type": "integer", "description": "Amount in kobo", "example": 4500000000 },
          "negotiable": { "type": "boolean", "example": true },
          "titleDocsVerified": { "type": "boolean", "example": false }
        }
      },
      "ShortstayDetails": {
        "type": "object",
        "required": ["nightlyRate"],
        "properties": {
          "nightlyRate": { "type": "integer", "description": "Amount in kobo", "example": 6500000 },
          "cleaningFee": { "type": "integer", "example": 1500000 },
          "minNights": { "type": "integer", "example": 1 },
          "maxNights": { "type": "integer", "example": 30 },
          "maxGuests": { "type": "integer", "example": 4 },
          "houseRules": { "type": "string", "nullable": true }
        }
      },
      "Listing": {
        "type": "object",
        "required": ["id", "ownerId", "listingType", "status", "title", "description", "propertyType", "bedrooms", "bathrooms", "address", "city", "state", "lat", "lng", "photos", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "ownerId": { "type": "string", "format": "uuid" },
          "listingType": { "type": "string", "enum": ["rent", "sale", "shortstay"] },
          "status": { "type": "string", "enum": ["draft", "active", "paused", "rented", "sold"] },
          "title": { "type": "string", "example": "3-bed flat in Lekki Phase 1" },
          "description": { "type": "string" },
          "propertyType": { "type": "string", "example": "apartment" },
          "bedrooms": { "type": "integer", "example": 3 },
          "bathrooms": { "type": "integer", "example": 2 },
          "areaSqm": { "type": "integer", "nullable": true, "example": 120 },
          "amenities": { "type": "array", "items": { "type": "string" }, "example": ["parking", "generator"] },
          "address": { "type": "string", "example": "12 Admiralty Way" },
          "city": { "type": "string", "example": "Lagos" },
          "state": { "type": "string", "example": "Lagos" },
          "lat": { "type": "number", "example": 6.4474 },
          "lng": { "type": "number", "example": 3.4700 },
          "photos": { "type": "array", "items": { "$ref": "#/components/schemas/ListingPhoto" } },
          "rent": { "$ref": "#/components/schemas/RentDetails", "nullable": true },
          "sale": { "$ref": "#/components/schemas/SaleDetails", "nullable": true },
          "shortstay": { "$ref": "#/components/schemas/ShortstayDetails", "nullable": true },
          "owner": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": { "type": "string" },
              "name": { "type": "string" },
              "avatarUrl": { "type": "string", "nullable": true },
              "listerType": { "type": "string", "nullable": true, "enum": ["agent", "landlord", null] }
            }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "UpdateListingRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 3 },
          "description": { "type": "string", "minLength": 10 },
          "amenities": { "type": "array", "items": { "type": "string" } },
          "bedrooms": { "type": "integer", "minimum": 0 },
          "bathrooms": { "type": "integer", "minimum": 0 }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/listings/{id}": {
      "get": {
        "tags": ["Listings"],
        "summary": "Get listing",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Listing detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Listing" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "patch": {
        "tags": ["Listings"],
        "summary": "Update listing",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateListingRequest" } } }
        },
        "responses": {
          "200": { "description": "Updated listing", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Listing" } } } },
          "403": { "description": "Not the listing owner", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Listings"],
        "summary": "Delete listing",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "204": { "description": "Deleted" },
          "403": { "description": "Not the listing owner", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}